Wholesale changes to preserve cookies in all cases
This wil make the tests way more stable from system to system.  We can't rely on the session cookie's name being the same between machines.  New #maintain_login function checks to see if there are cookies set and if there are not, we login again.  We use this for all but the login scripts .. those are the only scripts we don't want to maintain the sessoin because we are testing whether it works.