Fix this test spec
This was passing previously only because I don't think it was actually switching to the test_user all the time.  I think it was probably landing on the admin user in most cases.

Now that we're maintaining login on all set user type changes, this can't happen and the tests seem like they're running consistently for the first time.

(They were previously failing when I turned video capture on.)