Implement ability to lock out DB seed on subsequent runs
This should only be used in development, but for times when you want to favor speed over data, you can use this 'skip_db_seed' key in the mysql array of cypress.env.json.

If you need the database populated again, it's as simple as deleting the .lock file to get it back.