Added step definitions for enabling/disabling surveys & longitudinal mode via the Project Setup page
Also added a custom parameter type "toggleAction" (enable or disable) which can (and probably should) replace other parameter types in other existing and future step definitions.

New toggling surveys step definition is partly redundant, as there is an old step definition for enabling surveys, but with some differences: 1) takes a PID parameter, 2) visits the page, and 3) causes test failure if surveys are already enabled. This new step definition allows disabling or enabling, and does not fail if the state is already as desired, instead logging a warning to the Cypress test runner logs. This should probably be revisited; do we want the test to fail? If not, do we want the warning to be "louder", since it can easily be lost in the test logs?