Update Cypress to 11.2.0
This might be the maximum version we can upgrade to until we determine whether session handling is something we can use moving forward.  Last time I looked at it, it wasn't mature enough to actually use.  Legacy cookie handling is removed beyond this version so we cannot upgrade beyond this without structural changes to RCTF.